+++ src/main/java/me/alfie/immersiveenchanting/api/datapack/DatapackKeys.java
@@ -7,7 +7,9 @@ import net.minecraft.resources.Identifier;
 
 import java.lang.invoke.CallSite;
 
+/**Internal DatapackKeys*/
 public final class DatapackKeys {
+
     public static final DatapackKey<CostRegistry> COST = new DatapackKey<>(ImmersiveEnchanting.MODID, "enchantment_costs");
     public static final DatapackKey<NodeSoundMap> NODE_SOUNDS = new DatapackKey<>(ImmersiveEnchanting.MODID, "sounds");
 }
