+++ src/main/java/me/alfie/immersiveenchanting/datapack/enchantment_cost/CostRegistry.java
@@ -3,8 +3,7 @@ package me.alfie.immersiveenchanting.datapack.enchantment_cost;
 import me.alfie.immersiveenchanting.ImmersiveEnchanting;
 import me.alfie.immersiveenchanting.api.datapack.DatapackKeys;
 import me.alfie.immersiveenchanting.datapack.enchantment_cost.codec.CostData;
-import me.alfie.immersiveenchanting.datapack.manager.ClientDatapackManager;
-import me.alfie.immersiveenchanting.datapack.manager.ServerDatapackManager;
+import me.alfie.immersiveenchanting.api.datapack.manager.ServerDatapackManager;
 import net.minecraft.core.Holder;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.Registries;
