+++ src/main/java/me/alfie/immersiveenchanting/event/ModEvents.java
@@ -3,8 +3,8 @@ package me.alfie.immersiveenchanting.event;
 import me.alfie.immersiveenchanting.api.ApiPostEvents;
 import me.alfie.immersiveenchanting.api.enchanting_tab.NodeClickHandlerRegistry;
 import me.alfie.immersiveenchanting.command.ModCommands;
-import me.alfie.immersiveenchanting.datapack.manager.ClientDatapackManager;
-import me.alfie.immersiveenchanting.datapack.manager.ServerDatapackManager;
+import me.alfie.immersiveenchanting.api.datapack.manager.ClientDatapackManager;
+import me.alfie.immersiveenchanting.api.datapack.manager.ServerDatapackManager;
 import me.alfie.immersiveenchanting.creativetab.ModCreativeTab;
 import me.alfie.immersiveenchanting.api.description.TooltipDescriptionExtensions;
 import me.alfie.immersiveenchanting.datapack.enchantment_cost.CostDatapack;
