+++ src/main/java/me/alfie/immersiveenchanting/gui/tab/enchanting/node/BranchManager.java
@@ -1,7 +1,6 @@
 package me.alfie.immersiveenchanting.gui.tab.enchanting.node;
 
 import me.alfie.immersiveenchanting.datapack.enchantment_cost.CostRegistry;
-import me.alfie.immersiveenchanting.datapack.manager.ClientDatapackManager;
 import me.alfie.immersiveenchanting.gui.EnchantingTableScreen;
 import net.minecraft.world.item.ItemStack;
